Rename test to basic
