+++ managedservices/basic/error.go
@@ -1,4 +1,4 @@
-package helm
+package basic
 
 import "github.com/giantswarm/microerror"
